Move EntityPathPart to own file
